+++ ddl/mock.go
@@ -19,6 +19,8 @@ import (
 
 	"github.com/coreos/etcd/clientv3"
 	"github.com/juju/errors"
+	"github.com/pingcap/tidb/ast"
+	"github.com/pingcap/tidb/context"
 	"github.com/pingcap/tidb/model"
 	goctx "golang.org/x/net/context"
 )
@@ -118,4 +120,18 @@ func (dr *mockDelRange) start() {
 // clear implements delRangeManager interface.
 func (dr *mockDelRange) clear() {
 	return
+}
+
+// MockTableInfo mocks a table info by create table stmt ast and a specified table id.
+func MockTableInfo(ctx context.Context, stmt *ast.CreateTableStmt, tableID int64) (*model.TableInfo, error) {
+	cols, newConstraints, err := buildColumnsAndConstraints(ctx, stmt.Cols, stmt.Constraints)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	tbl, err := buildTableInfo(nil, stmt.Table.Name, cols, newConstraints, ctx)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	tbl.ID = tableID
+	return tbl, nil
 }
